refactor: Only use Mutex for render buffer
I don't really like this whole immediate GUI thing, but the easiest
thing so far seems to call the render function directly from the ui
closure. This way, the gui only needs to keep a pointer to the render
buffer, not to the whole application state, which then doesn't need to
be wrapped in a pointer at all to be used as a mutable reference in the
main event loop. With this commit, the GPU framebuffer is still not
updated, but at least when you click render the render buffer in memory
is actually updated, which can be seen by writing a sample file on disk.